feat: setup unit tests with PHPUnit
