+++ composer.json
@@ -3,5 +3,8 @@
         "sentry/sdk": "2.1.0",
         "php-http/curl-client": "^2.1",
         "swaggest/json-schema": "^0.12.29"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^9"
     }
 }
