+++ schemas/keyboard_info.distribution/1.0.6/keyboard_info.distribution.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "type": "object",
+  "allOf": [
+    { "$ref": "/keyboard_info.source/1.0.6/keyboard_info.source.json#/definitions/KeyboardInfo" },
+    { "required": [ "id", "name", "license", "languages", "lastModifiedDate", "platformSupport" ] }
+  ]
+}
