+++ schemas/search/1.0.2/search.json
@@ -1,18 +1,18 @@
-{ 
+{
   "$schema": "http://json-schema.org/schema#",
   "$ref": "#/definitions/Search",
-  
+
   "definitions": {
     "Search": {
       "type": "object",
       "properties": {
         "languages": { "type": "array", "items": { "$ref": "#/definitions/SearchLanguage" } },
-        "keyboards": { "type": "array", "items": { "$ref": "/keyboard_info.source/1.0.5/keyboard_info.source.json#/definitions/KeyboardInfo" } },
+        "keyboards": { "type": "array", "items": { "$ref": "/keyboard_info.source/1.0.6/keyboard_info.source.json#/definitions/KeyboardInfo" } },
         "countries": { "type": "array", "items": { "$ref": "#/definitions/SearchCountry" } },
         "rangematch": { "type": "string" }
       }
     },
-    
+
     "SearchCountry": {
       "type": "object",
       "properties": {
@@ -22,7 +22,7 @@
         "languages": { "type": "array", "items": { "$ref": "#/definitions/SearchLanguage" } }
       }
     },
-  
+
     "SearchLanguage": {
       "type": "object",
       "properties": {
