+++ tests/manual/.gitignore
