+++ tests/manual/langtags/package.json
