+++ tests/manual/langtags/validate.js
