+++ tests/manual/validate-kmw-api-schemas.sh
