+++ tests/manual/validate-kmw-api-transition.php
